Sketch out walking up path to find working directory
Behaves a-la git where the working directory is determined by walking up from
PWD until a .git directory is found.  Then the working directory is set as the
parent directory and, in the case of prototype, the local directory is set as
.prototype within the working directory.

The end product of which is the ability to specify config/plugins that apply to
a subtree of directories.